Working buttons, potential changes in comments
I am currently working on making the login page the first page when the app is opened and navigating to the home page when clicking 'submit'. I am currently figuring out how to use Routes, but since they do not currently work, these changes are commented out to be worked on in the future.